requirements.txt
astroid==1.4.6
colorama==0.3.7
docopt==0.6.2
lazy-object-proxy==1.2.2
pylint==1.5.6
python-gitlab==0.12.2
pywin32==220
rcfile==0.1.4
requests==2.9.1
six==1.10.0
vboxapi==1.0
wrapt==1.10.8
